+++ src/payload.config.ts
@@ -2,14 +2,19 @@ import path from 'path';
 
 import { buildConfig } from 'payload/config';
 
+import Guests from './collections/Guests';
 import Pages from './collections/Pages';
+import Parties from './collections/Parties';
+import Relations from './collections/Relations';
+import Sides from './collections/Sides';
 import Users from './collections/Users';
 
 export default buildConfig({
   admin: {
     user: Users.slug,
+    css: path.resolve(__dirname, 'custom/styles/index.scss'),
   },
-  collections: [Pages, Users],
+  collections: [Guests, Pages, Parties, Relations, Sides, Users],
   typescript: {
     outputFile: path.resolve(__dirname, 'payload-types.ts'),
   },
@@ -17,4 +22,6 @@ export default buildConfig({
     schemaOutputFile: path.resolve(__dirname, 'generated-schema.graphql'),
   },
   cors: [process.env.MONGODB_IP].filter(Boolean),
+  csrf: [process.env.SERVER_URL, process.env.DOMAIN].filter(Boolean),
+  serverURL: process.env.SERVER_URL,
 });